+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - Single letter module names are now permitted.
-- Added support for bitstring syntax.
+- Added support for bit_string syntax.
 - Remove old list prepend syntax `[x | y]`, use `[x .. y]` instead.
 
 ## v0.9.0 - 2020-06-01